week 4/requirements.txt
groq
transformers
torch
rouge-score
nltk
python-dotenv
requests
